+++ README.md
@@ -8,41 +8,48 @@
 [![cmake version](https://img.shields.io/badge/cmake-3.20.1-black)](https://cmake.org/)
 [![cpp version](https://img.shields.io/badge/c++-20-black)]()
 
-
 More information is coming soon.
 
 ## Requirements
+
 #### Platform:
+
 - [Windows 10](https://www.microsoft.com/en-us/software-download/windows10)
 - [Ubuntu 22.04](https://ubuntu.com/download/desktop)
 
 ## Package Version
 
-
 ## Installation
 
 #### Windows
+
 ```bash
 git clone https://github.com/USI-FMAA/digital_casting_app.git
 git clone --recurse-submodules -j8 https://github.com/microsoft/vcpkg
+
 # install package
 .\VCPKG\bootstrap-vcpkg.bat
 
 # initialize and build app
 .\package-install.bat
-
 ```
 
 ## Development
 
 #### Windows
+
 ```bash
+# update the package manger
 git submodule add https://github.com/microsoft/vcpkg VCPKG
+# install package
 .\VCPKG\bootstrap-vcpkg.bat
+# initialize and build app
 .\package-install.bat
+# test app
+.\build\debug\dcs_app.exe
 ```
 
-
 ## Credits
+
 This package was created by [WeiTing Chen](https://github.com/WeiTing1991)
 at [USI-FMAA](https://github.com/USI-FMAA) and [ETHZurich DFab](https://dfab.ch/).
