+++ tests/boost/add.py
@@ -1,0 +1,2 @@
+def add(i, j):
+    return i + j
